--- apps/web/app/lib/types/pipelines.ts
@@ -1,3 +1,0 @@
-/* v8 ignore start */
-export type { TResponsePipelineJobData as TPipelineInput } from "@formbricks/jobs";
-/* v8 ignore stop */
